implement Discott saturation Algorithm
This commit implements a new saturation algorithm called Discott.

It is based on the idea to keep the invariant that all clauses in
Active and Passive are always simplified by all clauses in Active.

The motivation behind this design choice is to keep the speed of
Discount (due to not performing simplifications between two clauses
in Passive), while keeping clauses in Passive simplified by clauses in
Active so that clause selection can use accurate weights (that is, it
can use the weight of each clause after all simplifications with Active
clauses have been performed).